# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  title: string;
  area: string;
  city: string;
  price: number;
  priceFormatted: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  sizeSqFt: number;
  imageUrl: string;
  description: string;
  features: string[];
  isFeatured: boolean;
  phase?: string;
  createdAt?: string;
}

export type PropertyStatsCityBreakdownItem = {
  city: string;
  count: number;
};

export type PropertyStatsTypeBreakdownItem = {
  type: string;
  count: number;
};

export interface PropertyStats {
  totalListings: number;
  avgPrice: number;
  cityBreakdown: PropertyStatsCityBreakdownItem[];
  typeBreakdown: PropertyStatsTypeBreakdownItem[];
}

export interface ChatMessageBody {
  message: string;
  sessionId: string;
  budget?: number;
  area?: string;
  city?: string;
  propertyType?: string;
  bedrooms?: number;
}

export type ChatResponseExtractedCriteria = {
  budget?: number;
  area?: string;
  city?: string;
  propertyType?: string;
  bedrooms?: number;
};

export interface ChatResponse {
  message: string;
  properties: Property[];
  sessionId: string;
  intent: string;
  extractedCriteria?: ChatResponseExtractedCriteria;
}

export interface ScheduleVisitRequest {
  propertyId: number;
  name: string;
  phone: string;
  email?: string;
  visitDate: string;
  visitTime: string;
  notes?: string;
}

export interface Visit {
  id: number;
  propertyId: number;
  property?: Property;
  name: string;
  phone: string;
  email?: string;
  visitDate: string;
  visitTime: string;
  notes?: string;
  status: string;
  createdAt?: string;
}

export type ListPropertiesParams = {
  city?: string;
  area?: string;
  minPrice?: number;
  maxPrice?: number;
  type?: string;
  bedrooms?: number;
};