# lib/api-zod/src/generated/types/chatMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageBody {
  message: string;
  sessionId: string;
  budget?: number;
  area?: string;
  city?: string;
  propertyType?: string;
  bedrooms?: number;
}