# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatResponseExtractedCriteria } from "./chatResponseExtractedCriteria";
import type { Property } from "./property";

export interface ChatResponse {
  message: string;
  properties: Property[];
  sessionId: string;
  intent: string;
  extractedCriteria?: ChatResponseExtractedCriteria;
}