# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  title: string;
  area: string;
  city: string;
  price: number;
  priceFormatted: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  sizeSqFt: number;
  imageUrl: string;
  description: string;
  features: string[];
  isFeatured: boolean;
  phase?: string;
  createdAt?: string;
}