# lib/api-zod/src/generated/types/propertyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatsCityBreakdownItem } from "./propertyStatsCityBreakdownItem";
import type { PropertyStatsTypeBreakdownItem } from "./propertyStatsTypeBreakdownItem";

export interface PropertyStats {
  totalListings: number;
  avgPrice: number;
  cityBreakdown: PropertyStatsCityBreakdownItem[];
  typeBreakdown: PropertyStatsTypeBreakdownItem[];
}