# lib/api-zod/src/generated/types/scheduleVisitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleVisitRequest {
  propertyId: number;
  name: string;
  phone: string;
  email?: string;
  visitDate: string;
  visitTime: string;
  notes?: string;
}