# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Property Matchmaker API
 * OpenAPI spec version: 0.1.0
 */
import type { Property } from "./property";

export interface Visit {
  id: number;
  propertyId: number;
  property?: Property;
  name: string;
  phone: string;
  email?: string;
  visitDate: string;
  visitTime: string;
  notes?: string;
  status: string;
  createdAt?: string;
}